Create Pool refactoring
