+++ packages/abacus-pools/src/sections/CreatePool/CreatePool.styled.tsx
